rivalz_demo/requirements.txt
rivalz-client
python-dotenv 